requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.1
django-admin-interface==0.26.1
django-adminlte-3==0.1.6
django-bootstrap4==23.2
django-colorfield==0.10.1
django-crispy-forms==2.0
django-froala-editor==4.1.4
django-grappelli==3.0.8
django-jazzmin==2.6.0
django-rest-framework==0.1.0
django-suit==0.2.28
djangorestframework==3.14.0
docutils==0.20.1
drf-yasg==1.21.7
filelock==3.12.4
fontawesomefree==6.4.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
Kivy==2.2.1
Kivy-Garden==0.1.5
kivymd==1.1.1

packaging==23.2
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
pycryptodome==3.19.0
Pygments==2.16.1
pytest==7.4.3
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
